Pretty bash prompt
The new bash PS1 is inspired by the vim powerline.

The first segment is the id of the current command in the history. The
second segment is the user and host, this segment is grey when the shell
is local and blue when it exists through ssh. The third segment is the
work directory. The last segment appears only when the current directory
is a git, hg or svn repositoty and display the current branch.